test(proptest): keep untranslated specs in their themed module
