+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 ### Added
 - **The `ONNX` column now measures the operators we *cannot* translate, not just the ones we can.** Specs only ever existed where `torch_to_nnef` succeeds, so the measured population was a strict subset of our own supported set: of the 171 operators the page marks unsupported, **zero** had ever been measured, and an operator ONNX handles and we do not could only be inherited as an unverified `✅*` claim from the retired TorchScript listing, never graded. The comparison was structurally blind exactly where we are weakest, and `Gap vs ONNX` was capped at the 20 rows that listing happened to name. `tests/proptest/op_specs/gaps/` closes it with **140 specs for operators that have no emitter at all**, grouped by why they are out of reach (`linalg` is one missing factorization repeated 33 times, `special` is 29 series evaluations, `rng` is a format question rather than a kernel one, `dataflow` splits data-dependent extents from memory layouts NNEF cannot name). The remaining 31 unsupported rows are listed in `EXCLUDED` with a reason each, and `test_every_unsupported_row_is_specced_or_excluded` fails if a row appears in neither, so the page and the catalog cannot drift apart. Measured result: **ONNX exports 43 of the 140**, 91 are gaps for both exporters, 5 are blocked before ONNX sees them. `Untested, no data` drops from 193 rows to 72, `Gap vs ONNX` grows from 20 to 44 and is now evidence rather than hearsay, and the `ONNX` bars move to `134/139` core and `394/581` overall. Two claims the old listing made are retracted by measurement (`masked_select`, `rrelu`: ONNX cannot export them either).
 - **`OpSpec.nnef_gap` marks a translation we do not ship, and is asserted rather than trusted.** A marker that only *skipped* would keep reporting an operator as unsupported long after someone implemented it, so the tract driver attempts the real export and checks it fails at the declared `stage`; both "it works now" and "it fails somewhere else" fail the test, with the fix named in the message. It is cheap, since a missing emitter raises during translation before tract is invoked. `tests/test_proptest_nnef_gap.py` adds the same guarantee to the **default** suite by cross-checking every `no-emitter` gap against the live emitter registry. `stage` distinguishes `no-emitter` from `export-error` (nine operators, mostly RNG factories, die in constant folding before the lookup) and from `raw-error` (`index_reduce` and `chalf` crash with a bare `TypeError` / `KeyError` instead of a `T2NError` naming the operator, which is a bug on our side either way).
+- **Untranslated specs live in the themed module they belong to**, not in a package of their own, so implementing an operator means deleting one field rather than moving a spec between files. `linalg`, `special` and `random_sampling` become themed modules in their own right (every operator in them is untranslated today, which is a property of the specs rather than of the module), and the rest join `shape`, `reductions`, `elementwise`, `factory`, `conv_pool`, `loss` and `specialty` next to the operators they are related to. Spec ids carry no marker of the gap either, so closing one does not churn pytest ids, hypothesis's example database or the artifact. Rows that get no spec at all are recorded in `op_specs/untranslated.py`.
 - **`OpSpec.nondeterministic`** skips only the numerics axis for RNG operators. Export and runtime stay measured; comparing two independent draws would report the definition of the operator rather than anything about the exporter.
 
 ### Changed
+- **The support page can no longer go stale unnoticed.** `test_page_credits_every_registered_emitter` compares the committed page against the live emitter registry, so adding an emitter without regenerating the page fails in the fast suite with the exact command to run. It needs no network, since it reads the registry rather than regenerating. Together with the two `EXCLUDED` checks, the page is now self-policing in both directions.
 - **`T2NErrorMissingOpEmitter`** now marks the one case that means "no translation exists at all", raised from `OpRegistry.get`. Its parent `T2NErrorNotImplemented` is also raised by emitters that exist but refuse a dtype, rank or attribute combination, so the two were indistinguishable to any caller trying to tell a missing operator from a partly-supported one. Subclasses the old exception, so existing handlers are unaffected. `add_new_aten_op.md` step 2 quotes the narrow class, and points contributors at the `Gap vs ONNX` filter for picking what to implement.
 
 ## [0.24.5] - 2026-07-28
