+++ tests/proptest/op_specs/__init__.py
@@ -6,11 +6,11 @@ submodules (elementwise, reductions, shape, activation, norm, conv_pool,
 specialty, factory); each exposes a single `SPECS` tuple and this module
 just concatenates them.
 
-`gaps` is the one module that does not describe something we ship: its
-specs carry `nnef_gap` and exist so the ONNX sweep can measure operators
-t2n cannot translate. They are part of the same registry on purpose, so
-every consumer (the sweep, the attribution check, the page) sees them
-without opting in.
+Operators we cannot translate live in these same themed modules, marked
+with `nnef_gap`, rather than in a package of their own. That way
+implementing one means deleting a field instead of moving a spec between
+files, and a spec's home never depends on its support status. The rows
+that get no spec at all are recorded in `untranslated.EXCLUDED`.
 """
 
 import typing as T
@@ -20,11 +20,13 @@ from . import (
     conv_pool,
     elementwise,
     factory,
-    gaps,
+    linalg,
     loss,
     norm,
+    random_sampling,
     reductions,
     shape,
+    special,
     specialty,
 )
 from ._common import NnefGap, NnefGapStage, OpSample, OpSpec
@@ -49,7 +51,9 @@ _MODULES = (
     specialty,
     factory,
     loss,
-    gaps,
+    linalg,
+    special,
+    random_sampling,
 )
 
 REGISTRY: T.Tuple[OpSpec, ...] = tuple(
